сброс выбранных офееров при смене типа
