+++ src/const.js
@@ -1,9 +1,25 @@
+import { offerType } from './mock/offer';
+
 const FilterType = {
   EVERYTHING: 'everything',
   FUTURE: 'future',
   PAST: 'past',
 };
 
+const BLANK_POINT = {
+  basePrice: 0,
+  dateFrom: new Date(),
+  dateTo: new Date(),
+  destination: {
+    name: '',
+    description: '',
+    pictures: []
+  },
+  id: 0,
+  isFavorite: false,
+  offers: [],
+  type: offerType[0],
+};
 const SortType = {
   DAY: 'day',
   EVENT: 'event',
@@ -54,4 +70,4 @@ const UpdateType = {
   MAJOR: 'MAJOR',
 };
 
-export { FilterType, SortTypes, SortType, UserAction, UpdateType };
+export { FilterType, SortTypes, SortType, UserAction, UpdateType, BLANK_POINT };
